+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+-   [annotation] factory `createConcentricSymbol` to create custom symbols
 -   [annotation] interactivity props to `BlockArrow`
 -   [core] component `TooltipDataComponent` to set tooltip data
 -   [core] prop `anchor` for `Label`